refactor: Split Swift files
- main.swift -> main, Command, Message
